Feature #11 added, Config_TLS.txt file now hold tls enabled configs
Issue number #11 requested a TLS enabled only config file, so i added a statement that uses the already created lines variable and check wether it contains "security=tls" in the string, if so that line will get written to the Config_TLS.txt file, i tried to not touch other parts of the code but i did notice something, line 104:
###
    elif os.path.exists(filename1):
        os.remove(filename1)
###
if the idea is to remove the files, the base 64 file should also be removed, but the elif statement checks if All_Config_Sub.txt exists, if so the code removes that file meaning the base64 file is never removed, i tested the code and it seemed to work tho i had some problem with unicode characters, they were being replaced with <undefined> so i had to run the code with python -X utf8 app.py, not a conceren if the code works on github actions. Please test for yourself and merge the request.